+++ ja2lib/Sector.h
@@ -20,7 +20,7 @@ UINT8 NumHostilesInSector(u8 sSectorX, u8 sSectorY, INT16 sSectorZ);
 
 // Returns TRUE if sector is under player control, has no enemies in it, and isn't currently in
 // combat mode
-BOOLEAN SectorOursAndPeaceful(INT16 sMapX, INT16 sMapY, INT8 bMapZ);
+BOOLEAN SectorOursAndPeaceful(u8 sMapX, u8 sMapY, INT8 bMapZ);
 
 // This will get an ID string like A9- OMERTA...
 void GetSectorIDString(u8 sSectorX, u8 sSectorY, INT8 bSectorZ, CHAR16* zString, size_t bufSize,
